backend/requirements.txt
fastapi
uvicorn
sqlalchemy
torch
torchvision
pillow
python-multipart
numpy
requests